+++ shape/design/activity_diagram.go
@@ -21,12 +21,15 @@ func (diagram *ActivityDiagram) LinkAll(s ...shape.Shape) {
 	}
 }
 
-// Link places an arrow with a label above it between the two shapes.
-func (diagram *ActivityDiagram) Link(from, to shape.Shape, txt string) {
+// Link places an arrow with a optional label above it between the two shapes.
+func (diagram *ActivityDiagram) Link(from, to shape.Shape, txt ...string) *shape.Arrow {
 	lnk := shape.NewArrowBetween(from, to)
 	diagram.Place(lnk)
-	lnk.SetClass("activity-arrow")
-	label := shape.NewLabel(txt)
-	diagram.Place(label).Above(lnk, 20)
-	diagram.VAlignCenter(lnk, label)
+	if len(txt) > 0 {
+		lnk.SetClass("activity-arrow")
+		label := shape.NewLabel(txt[0])
+		diagram.Place(label).Above(lnk, 20)
+		diagram.VAlignCenter(lnk, label)
+	}
+	return lnk
 }